+++ controller/settingscontroller.php
@@ -22,20 +22,37 @@
  * in every copy of the program you distribute. 
  * Pursuant to Section 7 § 3(e) we decline to grant you any rights under trademark law for use of our trademarks.
  *
-*/
+ */
 
 namespace OCA\Onlyoffice\Controller;
 
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http\TemplateResponse;
+use OCP\IL10N;
+use OCP\ILogger;
 use OCP\IRequest;
 
 use OCA\Onlyoffice\AppConfig;
+use OCA\Onlyoffice\DocumentService;
 
 /**
  * Settings controller for the administration page
  */
 class SettingsController extends Controller {
+
+    /**
+     * l10n service
+     *
+     * @var IL10N
+     */
+    private $trans;
+
+    /**
+     * Logger
+     *
+     * @var ILogger
+     */
+    private $logger;
 
     /**
      * Application configuration
@@ -45,16 +62,22 @@ class SettingsController extends Controller {
     private $config;
 
     /**
-     * @param string $AppName application name
-     * @param IRequest $request request object
-     * @param OCA\Onlyoffice\AppConfig $config application configuration
+     * @param string $AppName - application name
+     * @param IRequest $request - request object
+     * @param IL10N $trans - l10n service
+     * @param ILogger $logger - logger
+     * @param OCA\Onlyoffice\AppConfig $config - application configuration
      */
-    public function __construct($AppName, 
+    public function __construct($AppName,
                                     IRequest $request,
+                                    IL10N $trans,
+                                    ILogger $logger,
                                     AppConfig $config
                                     ) {
         parent::__construct($AppName, $request);
 
+        $this->trans = $trans;
+        $this->logger = $logger;
         $this->config = $config;
     }
 
@@ -64,20 +87,34 @@ class SettingsController extends Controller {
      * @return TemplateResponse
      */
     public function index() {
-        $data = ["documentserver" => $this->config->GetDocumentServerUrl()];
+        $data = [
+            "documentserver" => $this->config->GetDocumentServerUrl(),
+            "secret" => $this->config->GetDocumentServerSecret()
+        ];
         return new TemplateResponse($this->appName, "settings", $data, "blank");
     }
 
     /**
      * Save the document server address
      *
-     * @param string $documentserver application name
+     * @param string $documentserver - document service address
+     * @param string $secret - secret key for signature
      *
      * @return array
      */
-    public function settings($documentserver) {
+    public function settings($documentserver, $secret) {
         $this->config->SetDocumentServerUrl($documentserver);
-        return ["documentserver" => $this->config->GetDocumentServerUrl()];
+        $this->config->SetDocumentServerSecret($secret);
+
+        $documentserver = $this->config->GetDocumentServerUrl();
+        if (!empty($documentserver)) {
+            $error = $this->checkDocServiceUrl();
+        }
+
+        return [
+            "documentserver" => $this->config->GetDocumentServerUrl(),
+            "error" => $error
+            ];
     }
 
     /**
@@ -89,5 +126,34 @@ class SettingsController extends Controller {
      */
     public function formats(){
         return $this->config->formats;
+    }
+
+
+    /**
+     * Checking document service location
+     *
+     * @param string $documentServer - document service address
+     *
+     * @return string
+     */
+    private function checkDocServiceUrl() {
+
+        $documentService = new DocumentService($this->trans, $this->config);
+
+        try {
+            $commandResponse = $documentService->CommandRequest("version");
+
+            $this->logger->debug("CommandRequest on check: " . json_encode($commandResponse), array("app" => $this->appName));
+
+            $version = floatval($commandResponse->version);
+            if ($version < 4.2) {
+                throw new \Exception($this->trans->t("Not supported version"));
+            }
+        } catch (\Exception $e) {
+            $this->logger->error("CommandRequest on check error: " . $e->getMessage(), array("app" => $this->appName));
+            return $e->getMessage();
+        }
+
+        return "";
     }
 }