+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# Change Log
+
+## 1.0.2
+### Added
+- logging
+- checking Document Server address on save
+- checking version of onlyoffice
+- set language of editor
+
+### Changed
+- replace own Response class to OCP\AppFramework\Http class from core
+
+### Security
+- jwt signature for request to Document Server
+
 ## 1.0.1
 - fix exception when versions app is disabled
 - adding protocol to document server url